refactored logged in page
part of login handling being logged in is now in own component
